requirements.txt
streamlit
libsql-client
pandas
plotly
python-dotenv
streamlit-antd-components
bcrypt
pytest
